// src/typings/graphql.types.ts
export enum BookingStatus {
  ACTIVE = "ACTIVE",
  CANCELLED = "CANCELLED",
  COMPLETED = "COMPLETED",
  CONFIRMED = "CONFIRMED",
  RESERVED = "RESERVED",
  TERMINATED = "TERMINATED",
}

export enum FixedCountry {
  Japan = "Japan",
  Korea = "Korea",
  Singapore = "Singapore",
  Thailand = "Thailand",
}

export enum LoginTokenType {
  BEARER = "BEARER",
}

export enum OrderStatus {
  ASSIGNED = "ASSIGNED",
  CANCELLED = "CANCELLED",
  COMPLETED = "COMPLETED",
  CONFIRMED = "CONFIRMED",
  PENDING = "PENDING",
}